backend/requirements.txt
Flask
Flask-Cors
transformers
torch
groq
python-dotenv